fix: entities not required (VF-000) (#304)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-000**

### Brief description. What is this change?
Zoran noticed that if you send an intent request with no `entities` array in the `payload`. 
Technically it is required, but we want to make it robust as possible for the user.

Funky behavior happens because our typeguard requires array be identified: https://github.com/voiceflow/general-runtime/blob/01446bc108b9f67e65b208a889acfd28db269329/lib/services/runtime/types.ts#L13

![image](https://user-images.githubusercontent.com/5643574/158216621-1d742717-fe65-4f9c-93b8-a7fbc9c15583.png)

I've added a quick check to sanitize the incoming request, admittedly this isn't the best spot and we could have a dedicated adapter layer to fix incoming request. But that might be a bigger overhaul.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>