fix: handle cases around WS and DM key version lookup (VF-3982) (#363)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-3982**

### Brief description. What is this change?

The `resolveVersionAlias` needs to support looking up a version ID using an alias (`"development"`, or `"production"`). But this can only occur if using a dialog manager (DM) API key though, so we need to fail if a workspace key (WS) is trying to resolve an alias.

A DM key can default to the attached project's `devVersion` if a version ID is not provided. But with a WS key a version ID must always be provided.

### Related PRs

<!-- List related PRs against other branches -->

- https://github.com/voiceflow/creator-api/pull/1023
- https://github.com/voiceflow/general-runtime/pull/363

Co-authored-by: Zhi Heng (Brennan) Lu <brennan.lu314@gmail.com>
Co-authored-by: Tyler Stewart <tyler.stewart@voiceflow.com>