feat: add get publishing endpoint (CT-000) (#428)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-000**

### Brief description. What is this change?
Added a new endpoint: 
`GET /public/:projectID/publishing` - this fetches the `version.platformData.publishing` field of a PUBLIC project

because of the `resolvePublicProjectID` middleware, the project's `project.apiPrivacy` must be public to get this information. This is important to expose because it allows us to fetch project settings and metadata from the adapters/clients.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>