fix: nlc open slot threshold (PL-000) (#432)
the `openSlot` for NLC (untrained) setting means that a slot can be essentially anything:
![Screen Shot 2022-10-30 at 11 09 59 PM](https://user-images.githubusercontent.com/5643574/198926281-38b68f27-9ffe-452d-b3f5-8e3dc7431113.png)
In the example above, the user could say `whatever whatever not related` or `my size is whatever whatever not related` and still trigger the intent (if the model is not trained).

Today, the openSlot resolves a `undefined` confidence value. We are 100% confidence when it is a perfect regex match, but with open slots its really up in the air because we don't have a regex database for all slot types.

This causes a problem with the capture step with looks for a threshold before jumping to any random intent:
https://github.com/voiceflow/general-runtime/blob/master/lib/services/runtime/handlers/captureV2.ts#L13

If we set it to `0.5` for now that's fine.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>