feat: added transcript call authorized via api key (PL-363) (#463)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements PL-363**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

Adds a new endpoint that behaves similar to the old `controllers.public.createTranscript`, except that it is authorized via DM API key rather than by checking the `apiPrivacy` field on the `project` document.

The `apiPrivacy` field is somewhat tricky to set. The natural place where it should be set is either with a UX toggle or as a part of the channel publish. However,

1. UX toggle requires designs from Mike which is out-of-scope at the moment
2. A custom Teams publish job that sets `apiPrivacy` requires a lot of boilerplate to setup the job on `general-service`, which is far more effort and more error-prone than desired for this small change. 



Co-authored-by: Tyler Stewart <tyler.stewart@voiceflow.com>
Co-authored-by: Pedro Carvalho <pedro.matosdecarvalho@voiceflow.com>