feat: disable channel handlers (BUG-245) (#480)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements BUG-245**

### Brief description. What is this change?

The plan is to move alexa and google to running through the general runtime through adapters. We moved over the channel specific handlers to retain functionality, but that actually breaks these projects inside the prototype tool.

Disabling these channel specific handlers for now until we can focus more on the adapters and runtime merging. Ideally we shouldn't even need these.

Co-authored-by: Tyler Stewart <tyler.stewart@voiceflow.com>