fix: prediction failure due to unexpected teamID value (BUG-311) [bugfix] (#493)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements BUG-311**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

The `general-runtime` originally received the `teamID` as `number`. The original code then applied the type-cast `Number()` to get around type issues resulting from the `teamID` being typed as `string` even though it was returning a `number` as a value. 

A recent refactor broke this assumption and `general-runtime` now receives a `string` containing a hashstring value. This gets converted by `Number()` into the value `null`, which when passed into `nlu-gateway`, fails validation and breaks prediction.

The issue arose because `teamID` is an inconsistent value in our system and changing the format from hashed/unhashed values can have ramifications for the entire codebase. 

EDIT: Turns out the reason this issue was left undetected is because when you test using the Prototype Tool (which was the main way used to test this code), the `teamID` is returned as hashed `string`. Whereas, if you test it via DM API, it returns an unhashed `number`. So depending on if you contact the DM API and Prototype Tool, the type of `teamID` is either `number` or `string` (which doesn't make sense) but the type definition of `teamID` is recorded as `string` only.