fix(runtime): set as running state if no request (BUG-304) (#495)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements BUG-304**

### Brief description. What is this change?

When we start from a block in canvas we send a null request with a stack containing the node we want to start from.
The runtime would automatically start in a `REQUEST` state, which causes the capture step to think the request is the user's response to the capture step, causing a no match to trigger.
Instead, if we don't get a request from the user (`null`), then start the runtime in a `RUNNING` state instead. This tells the capture block that we're entering the block and to wait for the user reply instead.

Co-authored-by: Tyler Stewart <tyler.stewart@voiceflow.com>