feat: auto memory on ai steps [bugfix] (LLM-000) (#507)
There has been a significant refactor to the "transcript" system that we were using for AI assist - instead of parsing all the traces at the end of an interaction request and using that as the "transcript", we are now dynamically adding to the transcript on every independent speak,text, etc. step.

This is possible because of the utility function `addOutputTrace` which is used everywhere now.
We save up to 10 turns of the conversation, and that is now in `state.variables` instead of `state.storage`

I've also abstracted most of the AI API calls to a `fetchPrompt` function, that works with the new `chat` endpoint in ml-gateway.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>