fix: prediction failing if not training luis (PL-522) [bugfix] (#513)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements PL-522**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

We are running out of LUIS resources and it is not possible to create more because of LUIS deprecation. To mitigate this issue, we will disable the creation of LUIS applications for free plan users (who are using VFNLU as their main NLU at the moment) to reduce the volume of LUIS apps being created.

To accommodate the above fixes, we need to ensure that `general-runtime` is working even if no LUIS application is trained. However, our existing codebase is very LUIS centric, so `general-runtime` checks if `nlp.appID` and `nlp.resourceID` exists, but this is LUIS-specific data which is no longer created.

This PR simplifies the check on prediction to prevent it from breaking.