feat: support for alexa adapter (PL-30) (#475)
**Fixes or implements PL-30**

### Brief description. What is this change?

Changes needed for general-runtime in order to support alexa adapter.
Mostly is simplification/fixes, the only real change is adding a new property when sending a speak trace as response of an entity being unfulfilled.

This is important because the Alexa adapter must ignore this kind of prompts. In this case alexa NLU is the one handling dialog/slot filling so it will by itself push the prompt. We need a way to signal this on the trace so that Alexa adapter doesnt send the prompt.

### Related PRs

<!-- List related PRs against other branches -->

- https://github.com/voiceflow/platform/pull/346
- https://github.com/voiceflow/alexa-service/pull/318
- https://github.com/voiceflow/creator-app/pull/6773
- https://github.com/voiceflow/general-service/pull/506
- https://github.com/voiceflow/libs/pull/434