feat: add nlu inference endpoint (PL-551) (#558)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements PL-551**

### Brief description. What is this change?
Adds a new inference endpoint as requested by ML team.

Trying to add a proper route with params without breaking much of the runtime route flow.
Created new middleware to ensure all params are set also as headers so that compatibility with older middlewares is guaranteed. Hopefully in the future we move out of this old route patterns and can set proper route standards.