feat: add moderation to llm calls (PL-553) (#588)
**Fixes or implements PL-553**

### Brief description. What is this change?
We need to add a LLM Moderation call to OpenAI for all models (even the ones that are not from OpenAI). This means that the OpenAI client must always be there for models.

Added a new Unleah client

Tweaked the AI client to be a dynamic instead of static service so that it can support FF via unleash.

Created a new service AISynthesis that pulls some util fns into it so that it can call the new dynamix AI client.

Tweaked some other util fns to make proper use of these new client/service accordingly.