feat: moderation llm calls (PL-553) (#597)
**Fixes or implements PL-553**

### Brief description. What is this change?
We need to add a LLM Moderation call to OpenAI for all models (even the ones that are not from OpenAI). This means that the OpenAI client must always be there for models.

Added a new Unleah client

Tweaked the AI client to be a dynamic instead of static service so that it can support FF via unleash.

Created a new service AISynthesis that pulls some util fns into it so that it can call the new dynamix AI client.

Tweaked some other util fns to make proper use of these new client/service accordingly.

Tweaks also so that the openAI/anthorpic API clients are generated just once, but the AIModel instances are generated on the fly with a given context (workspace/project) so that context can be used during the model operations (as logging the projectID if the moderation call fails)

If no `OPENAI_API_KEY` is given content moderation is skipped completely because atm azure openai doesnt support moderation calls (imagine private clouds where they want any calls be made to "normal openai" they only trust thei own "azure openai").