fix: openai client layout (CT-000) (#603)
Overall the business need is to be able to turn off models based on ENV VAR configuration.
For example if no Anthropic API Key is given, runtime should still work, but any calls to "anthropic" will fail.
Today the whole thing will crash if any LLM env vars are not set.

So ultimately what is happening here is we are moving `AIModelContext` as part of `CompletionOptions` which is only triggered on calls to `getCompletion` or `getChatCompletion`.

So instead of injecting the `{ workspaceID, projectID }` as part of the GET to the model, it is now being passed when we make the actual calls. I think this makes sense because the context is part of the call and relevant to moderation, but has nothing to do with the model itself.

On the models side, we are no longer instantiating a new client every time someone calls `services.ai.get(model)` - but rather all models are instantiated at initalization and there is a fixed pool.
The clients are now living inside the model instead of being decoupled. 


Co-authored-by: Tyler Han <tylerhan97@gmail.com>