fix: add openai path (CV3-515) (#606)
Add a new (optional) `OPENAI_API_ENDPOINT` env var. If not set the client just uses the default OpenAI API so this is very low risk.

We can use the FF to disable moderation for a private cloud. Just moved it up higher. It's turned on for 100% of users right now.

Tested that things work:
<img width="513" alt="Screenshot 2023-10-18 at 6 23 22 PM" src="https://github.com/voiceflow/general-runtime/assets/5643574/170d5291-0593-4e16-be83-aa86de796cd0">

Co-authored-by: Tyler Han <tylerhan97@gmail.com>