fix: allow public origins (CT-000) (#616)
for the webchat we use these `/public/` endpoints that anyone can hit.

For private clouds, we can set an environment variable that restricts what origins requests can come in from.

Obviously one bad actor and spoof their origin in the request - but they can't propagate this to other unknowning users as long as they use a standard browser. 

I don't think this is the most beautiful implementation of a check ever, but for now this is very specific to only one client. We can re-evaluate this when there are more use cases

Co-authored-by: Tyler Han <tylerhan97@gmail.com>
Co-authored-by: Xavier Portilla Edo <xavi_tb@hotmail.com>