feat: pass intent & entity scopes (VF-000) (#640)
Determine intent/entity scope and pass it to nlu-gateway via:
- `filteredIntents: string[]`
- `filteredEntities: string[]`
- `excludeFilteredIntents: boolean`
- `excludeFilteredEntities: boolean`

The scope is defined by the intersection of:
- "intent scoping" at the `Listen` step level (`Buttons`, `Choice`, `Capture`) of the current node
- reachable intents determined by all commands in all frames on the runtime stack. ("Available from other topics" toggle on intent block)

Changes are backwards compatible and have no impact on functionality at this point. 

For context, these parameters will be passed to NLU user inference in order to better calculate accuracies based on a refined intent/entity scope rather than the full set of intents/entities.