feat: llm inference [bugfix] (CT-000) (#638)
add LLM inference to NLU.
For the sake of making sonarcloud happy about the complexity, I've moved a bunch of code segements out into individual functions.

I'll add more comments in the PR code explaining things.

l know that the logic in `lib/services/nlu/llmHybrid.ts` is quite verbose, but for the time being, this is an MVP and a lot of it is subject to change. I don't want to invest too much into cleaning up or streamlining patterns if there are potentially big changes.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>