fix: bypass NLU for local scoped captureV2 [bugfix] (CT-000) (#666)
Yoyo and I disabled the intents/entities scoping feature by @e-vandenberg, but this has resurfaced some new issues withe scoping.

Bypass NLU for captureV2 with local scoping. There is 0 reason to use the NLU in this configuration:
<img width="470" alt="Screenshot 2023-12-28 at 2 25 18 PM" src="https://github.com/voiceflow/general-runtime/assets/5643574/33664de2-6ce6-45e4-b270-9307efb8c464">

The problem right now is it might trigger entity filling if you match an intent.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>