fix: bump minor versions (BUG-1172) (#713)
Another attempt to address memory issues. 

We're noticing it creep up but looking through the heapdump it is still incredibly hard to pinpoint where the particular issue is. Trying to update a lot of dependencies slowly as well.

Reference my comments for the decisions behind each bump.
More major bumps I will do as separate and distinct PRs.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>
Co-authored-by: Xavier Portilla Edo <xavi_tb@hotmail.com>