fix: use immer 10 (PL-000) (#714)
We don't actually use immer much, but immer 10 has a 33% peformance upgrade and is generally nice.
https://github.com/immerjs/immer/releases/tag/v10.0.0

The only place in the code I could functionally trace us using it is to pop frames off the stack:
https://github.com/voiceflow/general-runtime/blob/7369f5037d7f083490a91501650f461813d4cdb5/runtime/lib/Runtime/Stack/index.ts#L48-L58

So pretty low risk, low blast radius. Just a side effect of some of my investigations.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>