fix: add dd-trace back to runtime (BUG-1172) (#721)
Pretty simple. dd-trace wasn't the source or the cause of the memory leak, but I do think it obfuscated and exasterbated the leak a lot.

Adding it back now. We should REALLY focus on upgrade to node 20 so we can use the latest version of dd-trace.

Reverting changes from:
https://github.com/voiceflow/general-runtime/pull/713
https://github.com/voiceflow/general-runtime/pull/712

We should still monitor memory usage carefully after this is deployed to be 100% sure that it wasn't the cause of another unrelated memory leak from https://github.com/voiceflow/general-runtime/pull/715

Co-authored-by: Tyler Han <tylerhan97@gmail.com>