fix: max node memory threshold [bugfix] (PL-000) (#728)
So our production instances are only at 1.5GB.

Under any regular circumstance, the node memory threshold should be higher than the pod's memory threshold, because this gives up visibility when something goes OOM.

Setting it to 4GB allows me to set the pod to just 4.5 GB for doing a heap dump. Otherwise I'd have to set it to 8.5GB or something for the 8GB configuration.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>