feat: don't call NLU if project.prototype.nlp not set (PL-949) (#787)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->
We allow publishing without training, so a model may not exist. Check `project.prototype.nlp` for existence of a model.

This is a temporary solution that needs a more robust check because it does not actually indicate that the current version is the one that is trained and available. But will lessen the log noise and unnecessary NLU calls

Co-authored-by: Greg O'Grady <gregory.ogrady@voiceflow.com>